+++ reactive/src/commonTest/kotlin/dev/toolkt/reactive/cell/test_utils/StatelessCellTestScenario.kt
@@ -2,8 +2,10 @@ package dev.toolkt.reactive.cell.test_utils
 
 import dev.toolkt.reactive.MomentContext
 import dev.toolkt.reactive.cell.Cell
-import dev.toolkt.reactive.event_stream.mapNotNull
-import dev.toolkt.reactive.event_stream.take
+import dev.toolkt.reactive.cell.observe
+import kotlin.test.assertEquals
+import kotlin.test.assertIs
+import kotlin.test.assertIsNot
 
 interface StatelessCellTestScenario {
     fun testObserved(
@@ -11,99 +13,140 @@ interface StatelessCellTestScenario {
     )
 }
 
-class ConfigurationContext(
-    private val momentContext: MomentContext,
-    private val ticker: Ticker,
-) {
-    private var isClosed = false
-
-    private var tMax = 0
-
-    fun recordTick(
-        tick: Tick,
+fun <InputT, ValueT> buildStatelessCellTestScenario(
+    configure: context (MomentContext, ConfigurationContext) () -> InputT,
+    instantiate: InputT.() -> Cell<ValueT>,
+    verificationTick: TickAlike,
+    expectedUpdatedValue: ValueT?,
+    shouldExpectFreeze: Boolean,
+): StatelessCellTestScenario = object : StatelessCellTestScenario {
+    override fun testObserved(
+        observationTick: TickAlike,
     ) {
-        require(!isClosed) {
-            "ConfigurationContext is already closed."
+        val properObservationTick = observationTick.asTick
+        val properVerificationTick = verificationTick.asTick
+
+        require(properObservationTick.t <= properVerificationTick.t) {
+            "Observation tick can't be later than expected update tick"
         }
 
-        if (tick.t > tMax) {
-            tMax = tick.t
-        }
-    }
-
-    fun getMomentContext(): MomentContext {
-        require(!isClosed) {
-            "ConfigurationContext is already closed."
+        if (properObservationTick == properVerificationTick) {
+            return // TODO: Implement pre-observed verification
         }
 
-        return momentContext
-    }
+        val ticker = Ticker()
 
-    fun getTicker(): Ticker {
-        require(!isClosed) {
-            "ConfigurationContext is already closed."
+        val (configuredInput, maxRecordedTick) = MomentContext.execute {
+            val configurationContext = ConfigurationContext(
+                momentContext = MomentContext.extract(),
+                ticker = ticker,
+            )
+
+            val configuredInput = with(configurationContext) {
+                configure()
+            }
+
+            val maxRecordedTick = configurationContext.getMaxRecordedTick()
+
+            configurationContext.close()
+
+            Pair(configuredInput, maxRecordedTick)
         }
 
-        return ticker
-    }
-
-    fun close() {
-        isClosed = true
-    }
-
-    fun getMaxRecordedTick(): Tick {
-        require(!isClosed) {
-            "ConfigurationContext is already closed."
+        require(properObservationTick.t <= maxRecordedTick.t) {
+            "Observation tick ${properObservationTick.t} exceeds the max recorded tick ${maxRecordedTick.t}"
         }
 
-        return Tick(t = tMax)
-    }
-}
+        require(properVerificationTick.t <= maxRecordedTick.t) {
+            "Verification tick ${properVerificationTick.t} exceeds the max recorded tick ${maxRecordedTick.t}"
+        }
 
-fun <ValueT : Any> defineConstCell(
-    constValue: ValueT,
-): Cell<ValueT> = Cell.of(constValue)
+        val subjectCell = configuredInput.instantiate()
 
-context(configurationContext: ConfigurationContext) fun <ValueT : Any> defineDynamicCell(
-    initialValue: ValueT,
-    updatedValueByTick: Map<TickAlike, ValueT>,
-    freezeTick: TickAlike?,
-): Cell<ValueT> {
-    val onTick = configurationContext.getTicker().onTick
-    val properFreezeTick = freezeTick?.asTick
-
-    updatedValueByTick.forEach { (tickAlike, _) ->
-        configurationContext.recordTick(
-            tick = tickAlike.asTick,
+        ticker.fastForward(
+            stopTick = properObservationTick,
         )
-    }
 
-    val updatedValueByProperTick = updatedValueByTick.entries.associate {  (tickAlike, updatedValue) ->
-        tickAlike.asTick to updatedValue
-    }
+        var receivedNotifications: MutableList<Cell.Notification<ValueT>>? = null
 
-    freezeTick?.let {
-        configurationContext.recordTick(
-            tick = it.asTick,
+        subjectCell.observe(
+            observer = object : Cell.Observer<ValueT> {
+                override fun handleNotification(
+                    notification: Cell.Notification<ValueT>,
+                ) {
+                    receivedNotifications?.add(notification)
+                }
+            },
         )
-    }
 
-    val onTickCropped = when (properFreezeTick) {
-        null -> onTick
+        ticker.fastForward(
+            stopTick = properVerificationTick,
+        )
 
-        else -> MomentContext.execute {
-            onTick.take(properFreezeTick.t + 1)
+        receivedNotifications = mutableListOf()
+
+        ticker.step(
+            properVerificationTick,
+        )
+
+        when {
+            expectedUpdatedValue == null && !shouldExpectFreeze -> {
+                val receivedNotificationCount = receivedNotifications.size
+
+                assertEquals(
+                    expected = 0,
+                    actual = receivedNotificationCount,
+                    message = "Expected no notifications at verification tick (got: $receivedNotificationCount)",
+                )
+            }
+
+            else -> {
+                assertEquals(
+                    expected = 1,
+                    actual = receivedNotifications.size,
+                    message = "Expected exactly one notification at verification tick",
+                )
+
+                val singleReceivedNotification = receivedNotifications.single()
+
+                when {
+                    expectedUpdatedValue != null -> {
+                        val singleReceivedUpdateNotification = assertIs<Cell.UpdateNotification<*>>(
+                            value = singleReceivedNotification,
+                            message = "Expected an update notification",
+                        )
+
+                        assertEquals(
+                            expected = expectedUpdatedValue,
+                            actual = singleReceivedUpdateNotification.updatedValue,
+                            message = "Received updated value differs from expected",
+                        )
+                    }
+
+                    else -> {
+                        assertIsNot<Cell.UpdateNotification<*>>(
+                            value = singleReceivedNotification,
+                            message = "Did not expect an update notification",
+                        )
+                    }
+                }
+
+                when {
+                    shouldExpectFreeze -> {
+                        assertIs<Cell.FreezeNotification<*>>(
+                            value = singleReceivedNotification,
+                            message = "Expected a freeze notification",
+                        )
+                    }
+
+                    else -> {
+                        assertIsNot<Cell.FreezeNotification<*>>(
+                            value = singleReceivedNotification,
+                            message = "Did not expect a freeze notification",
+                        )
+                    }
+                }
+            }
         }
-    }
-
-    val newValues = onTickCropped.mapNotNull { tick ->
-        updatedValueByProperTick[tick]
-    }
-
-    return with(configurationContext.getMomentContext()) {
-        Cell.define(
-            initialValue = initialValue,
-            newValues = newValues,
-        )
     }
 }
